Make map updating a function of the map object Fixes #121
