+++ js/function/update_pd_td_difference.js
@@ -1,17 +1,14 @@
-import pick_colour from "./pick_colour.js";
 import quantile from "./quantile.js";
-import highlight from "./highlight.js";
-import highlight_off from "./highlight_off.js";
 import regression_line from "./regression_line.js";
 
-export default function update_pd_td_difference(map_svg, path, plot) {
+export default function update_pd_td_difference(map, plot) {
     // This function requires both PD and (a subset of) TD to be retrieved.
     // Since these are obtained by processes effectively acting in parallel,
     // there isn't a suitable context where both TD and PD exist together.
     // To solve this issue, this function isn't passed the TD and PD data,
     // but instead retrieves it from what's already plotted on the map.
 
-    const td_data = map_svg
+    const td_data = map
         .select("g.td")
         .selectChildren()
         .data();
@@ -19,13 +16,13 @@ export default function update_pd_td_difference(map_svg, path, plot) {
     const td = d => d.properties.diversity;
 
     // Get PD from map squares with IDs matching the TD nodes
-    const pd = d => map_svg
+    const pd = d => map
         .select("g.pd")
         .select(`#pd_${d.key}`)
         .datum()
         .properties.pd;
 
-    const id = map_svg
+    const id = map
         .select("g.pd")
         .attr("id");
 
@@ -56,39 +53,15 @@ export default function update_pd_td_difference(map_svg, path, plot) {
         });
 
     const diff = f => f.properties.difference;
-    const max = d3.max(diff_features, f => Math.abs(diff(f)));
 
-    map_svg.selectAll(".data")
-        .attr("visibility", "hidden");
-    let group = map_svg.selectAll("g.diff");
-    if (group.empty())
-        group = map_svg.insert("g", "#borders")
-            .attr("class", "data diff")
-            .attr("id", `diff_${id}`);
-    group
-        .attr("visibility", "visible")
-        .selectAll("path")
-        .data(diff_features, d => d ? d.key : this.id.substring(5))
-        .join(
-            enter => enter
-                .append("path")
-                .attr("class", f => `highlightable key_${f.key}`)
-                .attr("id", f => `diff_${f.key}`)
-                .attr("fill", f =>
-                    pick_colour(f, diff_features, diff, "", 2))
-                .attr("fill-opacity", f =>
-                    0.5 * (1 + Math.abs(diff(f)) / Math.max(max, 1)))
-                .attr("d", path)
-                .text(f => f.text)
-                .on("mouseover", highlight)
-                .on("mouseleave", highlight_off),
-            update => update
-                .text(f => f.text)
-                .attr("fill", f =>
-                    pick_colour(f, diff_features, diff, "", 2))
-                .attr("fill-opacity", f =>
-                    0.5 * (1 + Math.abs(diff(f)) / Math.max(max, 1)))
-        );
+    map.data(
+        `diff_${id}`,
+        "diff",
+        diff_features,
+        d => d.key,
+        d => d ? d.key : this.id.substring(5),
+        diff,
+        d => d.text);
 
     plot.set_colours(diff_features, diff, d => d ? d.key : this.id.substring(5));
     plot.set_text(diff_features, f => f.text, d => d ? d.key : this.id.substring(5));
